+++ src/nacl/__init__.py
@@ -12,10 +12,6 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from __future__ import absolute_import, division, print_function
-
-import sys
-import warnings
 
 __all__ = [
     "__title__",
@@ -30,7 +26,7 @@ __all__ = [
 
 __title__ = "PyNaCl"
 __summary__ = (
-    "Python binding to the Networking and Cryptography (NaCl) " "library"
+    "Python binding to the Networking and Cryptography (NaCl) library"
 )
 __uri__ = "https://github.com/pyca/pynacl/"
 
@@ -40,14 +36,4 @@ __author__ = "The PyNaCl developers"
 __email__ = "cryptography-dev@python.org"
 
 __license__ = "Apache License 2.0"
-__copyright__ = "Copyright 2013-2018 {0}".format(__author__)
-
-
-if sys.version_info[0] == 2:
-    warnings.warn(
-        "Python 2 is no longer supported by the Python core team. Support for "
-        "it is now deprecated in PyNaCl, and will be removed in the "
-        "next release.",
-        DeprecationWarning,
-        stacklevel=2,
-    )
+__copyright__ = "Copyright 2013-2018 {}".format(__author__)