+++ src/nacl/bindings/randombytes.py
@@ -12,7 +12,6 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from __future__ import absolute_import, division, print_function
 
 from nacl import exceptions as exc
 from nacl._sodium import ffi, lib
@@ -44,7 +43,7 @@ def randombytes_buf_deterministic(size, seed):
     """
     if len(seed) != randombytes_SEEDBYTES:
         raise exc.TypeError(
-            "Deterministic random bytes must be generated " "from 32 bytes"
+            "Deterministic random bytes must be generated from 32 bytes"
         )
 
     buf = ffi.new("unsigned char[]", size)