+++ tests/test_signing.py
@@ -12,7 +12,6 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from __future__ import absolute_import, division, print_function
 
 import binascii
 
@@ -28,6 +27,8 @@ from nacl.bindings import (
 from nacl.encoding import Base64Encoder, HexEncoder
 from nacl.exceptions import BadSignatureError
 from nacl.signing import SignedMessage, SigningKey, VerifyKey
+
+from .utils import assert_equal, assert_not_equal, read_crypto_test_vectors
 
 
 def tohex(b):